# Improve performance for a few `Rows` methods